Use just rgblight_set instead of full custom driver
